# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para interagir com o Redirectify",
        "title": "API do Redirectify",
        "contact": {
            "name": "Equipe da DevOps (Pablo Eduardo, Guilherme Bernardo e Eduardo Henrique)",
            "email": "comercialdevops@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/redirecionadores": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadores"
                ],
                "summary": "Cria um redirecionador",
                "parameters": [
                    {
                        "description": "Nome",
                        "name": "nome",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Ordem de Redirecionamento",
                        "name": "ordem_de_redirecionamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/redirecionadores/:codigo_hash": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadors"
                ],
                "summary": "Retorna o redirecionador com o código hash fornecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Redirecionador"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadors"
                ],
                "summary": "Remove um redirecionador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadores"
                ],
                "summary": "Atualiza um redirecionador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nome",
                        "name": "nome",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Ordem de Redirecionamento",
                        "name": "ordem_de_redirecionamento",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/redirecionadores/:nome_de_usuario": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadores"
                ],
                "summary": "Retorna os redirecionadores",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Redirecionador"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/redirecionadores/rehash/:codigo_hash": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redirecionadores"
                ],
                "summary": "Recria o hash de um redirecionador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Erro": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "Redirecionador": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "codigo_hash": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "ordem_de_redirecionamento": {
                    "type": "string"
                },
                "removido_em": {
                    "type": "string"
                },
                "usuario": {
                    "type": "string"
                }
            }
        },
...
    }
}